requirements.txt
chainlit
ollama
langchain>=0.3.0
langchain-core>=0.3.0
langchain-community>=0.3.0
langchain-huggingface>=0.1.0
langchain-qdrant>=0.2.0
qdrant-client>=1.12.0
sentence-transformers>=3.0.0
FlagEmbedding>=1.2.0
rank_bm25>=0.2.2
jieba
json_repair
nltk
rouge
nano-vectordb
networkx
numpy
tiktoken
pipmaster
minirag-hku @ git+https://github.com/HKUDS/MiniRAG.git
